FIX: Check if path exists before removing in zip submissions

+++ app/models/unit.rb
@@ -939,11 +939,43 @@ class Unit < ActiveRecord::Base
   end
 
   #
+  # Create a temp zip file with all submission PDFs for a task
+  #
+  def get_task_submissions_pdf_zip(current_user, td)
+    # Get a temp file path
+    filename = FileHelper.sanitized_filename("submissions-#{code}-#{td.abbreviation}-#{current_user.username}-pdfs")
+    result = Tempfile.new([filename, '.zip'])
+
+    tasks_with_files = td.related_tasks_with_files
+
+    # Create a new zip
+    Zip::File.open(result.path, Zip::File::CREATE) do |zip|
+      Dir.mktmpdir do |dir|
+        # Extract all of the files...
+        tasks_with_files.each do |task|
+          path_part = if td.is_group_task? && task.group
+                        task.group.name.to_s
+                      else
+                        task.student.username.to_s
+                      end
+
+          FileUtils.cp task.portfolio_evidence, File.join(dir, path_part.to_s) + '.pdf'
+        end # each task
+
+        # Copy files into zip
+        zip_root_path = "#{td.abbreviation}-pdfs"
+        FileHelper.recursively_add_dir_to_zip(zip, dir, zip_root_path)
+      end # mktmpdir
+    end # zip
+    result
+  end
+
+  #
   # Create a temp zip file with all submissions for a task
   #
   def get_task_submissions_zip(current_user, td)
     # Get a temp file path
-    filename = FileHelper.sanitized_filename("submissions-#{code}-#{td.abbreviation}-#{current_user.username}.zip")
+    filename = FileHelper.sanitized_filename("submissions-#{code}-#{td.abbreviation}-#{current_user.username}-files.zip")
     result = Tempfile.new(filename)
 
     tasks_with_files = td.related_tasks_with_files
@@ -963,7 +995,7 @@ class Unit < ActiveRecord::Base
                                       ->(_task, to_path, name) { File.join(to_path.to_s, path_part, name.to_s) }) # call
 
           FileUtils.mv Dir.glob("#{dir}/#{path_part}/#{task.id}/*"), File.join(dir, path_part.to_s)
-          FileUtils.rm_r "#{dir}/#{path_part}/#{task.id}"
+          FileUtils.rm_r "#{dir}/#{path_part}/#{task.id}" if File.directory?("#{dir}/#{path_part}/#{task.id}")
         end # each task
 
         # Copy files into zip
